src/preprocessing/audio_processor.py
import librosa
import numpy as np
import soundfile as sf
from scipy import signal
import yaml

class AudioProcessor:
    def __init__(self, config_path='config.yaml'):
        # Audio feature extraction implementation
        pass

    def extract_audio_features(self, audio_file_path):
        """Extract audio features dari file audio"""
        # Load audio
        y, sr = librosa.load(audio_file_path, sr=self.config['data']['sample_rate'])
        
        # Extract MFCC
        mfcc = librosa.feature.mfcc(y=y, sr=sr, n_mfcc=self.config['data']['n_mfcc'])
        
        # Extract other features
        spectral_centroid = librosa.feature.spectral_centroid(y=y, sr=sr)
        spectral_rolloff = librosa.feature.spectral_rolloff(y=y, sr=sr)
        zero_crossing_rate = librosa.feature.zero_crossing_rate(y)
        
        # Combine features
        features = np.concatenate([
            np.mean(mfcc, axis=1),
            np.mean(spectral_centroid),
            np.mean(spectral_rolloff),
            np.mean(zero_crossing_rate)
        ])
        
        return features