requirements.txt
# Core ML Libraries
catboost==1.2.2
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
shap==0.42.1

# Audio Processing
librosa==0.10.1
scipy==1.11.1
soundfile==0.12.1

# Optimization
deap==1.3.3

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Web Framework (untuk nanti)
fastapi==0.101.1
uvicorn==0.23.2
streamlit==1.25.0

# Utilities
pyyaml==6.0.1
joblib==1.3.2
tqdm==4.65.0